Move client-related code back to @liveblocks/client
