+++ packages/liveblocks-core/src/__tests__/_utils.ts
@@ -28,7 +28,6 @@ import type {
   ToImmutable,
 } from "../types";
 import { ClientMsgCode, CrdtType, ServerMsgCode } from "../types";
-import { remove } from "../utils";
 import type { JsonStorageUpdate } from "./_updatesUtils";
 import { serializeUpdateToJson } from "./_updatesUtils";
 
@@ -49,94 +48,6 @@ function deepClone<T extends Json>(items: T): T {
   // won't lead to type unsafety, so this use case is okay.
   // eslint-disable-next-line no-restricted-syntax
   return JSON.parse(JSON.stringify(items));
-}
-
-export class MockWebSocket implements WebSocket {
-  CONNECTING = 0;
-  OPEN = 1;
-  CLOSING = 2;
-  CLOSED = 3;
-
-  static instances: MockWebSocket[] = [];
-
-  isMock = true;
-
-  callbacks = {
-    open: [] as Array<(event?: WebSocketEventMap["open"]) => void>,
-    close: [] as Array<(event?: WebSocketEventMap["close"]) => void>,
-    error: [] as Array<(event?: WebSocketEventMap["error"]) => void>,
-    message: [] as Array<(event?: WebSocketEventMap["message"]) => void>,
-  };
-
-  sentMessages: string[] = [];
-  readyState: number;
-
-  constructor(
-    public url: string,
-    private onSend: (message: string) => void = () => {}
-  ) {
-    this.readyState = this.CLOSED;
-    MockWebSocket.instances.push(this);
-  }
-
-  close(_code?: number, _reason?: string): void {
-    this.readyState = this.CLOSED;
-  }
-
-  addEventListener<T extends "open" | "close" | "error" | "message">(
-    event: T,
-    callback: (event: WebSocketEventMap[T]) => void
-  ) {
-    this.callbacks[event].push(callback as any);
-  }
-
-  removeEventListener<T extends "open" | "close" | "error" | "message">(
-    event: T,
-    callback: (event: WebSocketEventMap[T]) => void
-  ) {
-    remove(this.callbacks[event], callback as any);
-  }
-
-  send(message: string) {
-    this.sentMessages.push(message);
-    this.onSend(message);
-  }
-
-  open() {
-    this.readyState = this.OPEN;
-    for (const callback of this.callbacks.open) {
-      callback();
-    }
-  }
-
-  closeFromBackend(event?: CloseEvent) {
-    this.readyState = this.CLOSED;
-    for (const callback of this.callbacks.close) {
-      callback(event);
-    }
-  }
-
-  // Fields and methods below are not implemented
-
-  binaryType: BinaryType = "blob";
-  bufferedAmount: number = 0;
-  extensions: string = "";
-  onclose: ((this: WebSocket, ev: CloseEvent) => any) | null = () => {
-    throw new Error("NOT IMPLEMENTED");
-  };
-  onerror: ((this: WebSocket, ev: Event) => any) | null = () => {
-    throw new Error("NOT IMPLEMENTED");
-  };
-  onmessage: ((this: WebSocket, ev: MessageEvent<any>) => any) | null = () => {
-    throw new Error("NOT IMPLEMENTED");
-  };
-  onopen: ((this: WebSocket, ev: Event) => any) | null = () => {
-    throw new Error("NOT IMPLEMENTED");
-  };
-  protocol: string = "";
-  dispatchEvent(_event: Event): boolean {
-    throw new Error("Method not implemented.");
-  }
 }
 
 export const FIRST_POSITION = makePosition();
