+++ packages/liveblocks-core/src/types/index.ts
@@ -198,43 +198,6 @@ export type RoomInitializers<
   shouldInitiallyConnect?: boolean;
 }>;
 
-export type Client = {
-  /**
-   * Gets a room. Returns null if {@link Client.enter} has not been called previously.
-   *
-   * @param roomId The id of the room
-   */
-  getRoom<
-    TPresence extends JsonObject,
-    TStorage extends LsonObject = LsonObject,
-    TUserMeta extends BaseUserMeta = BaseUserMeta,
-    TRoomEvent extends Json = never
-  >(
-    roomId: string
-  ): Room<TPresence, TStorage, TUserMeta, TRoomEvent> | null;
-
-  /**
-   * Enters a room and returns it.
-   * @param roomId The id of the room
-   * @param options Optional. You can provide initializers for the Presence or Storage when entering the Room.
-   */
-  enter<
-    TPresence extends JsonObject,
-    TStorage extends LsonObject = LsonObject,
-    TUserMeta extends BaseUserMeta = BaseUserMeta,
-    TRoomEvent extends Json = never
-  >(
-    roomId: string,
-    options: RoomInitializers<TPresence, TStorage>
-  ): Room<TPresence, TStorage, TUserMeta, TRoomEvent>;
-
-  /**
-   * Leaves a room.
-   * @param roomId The id of the room
-   */
-  leave(roomId: string): void;
-};
-
 /**
  * Represents all the other users connected in the room. Treated as immutable.
  */
@@ -283,28 +246,6 @@ export type Polyfills = {
   fetch?: typeof fetch;
   WebSocket?: any;
 };
-
-/**
- * The authentication endpoint that is called to ensure that the current user has access to a room.
- * Can be an url or a callback if you need to add additional headers.
- */
-export type ClientOptions = {
-  throttle?: number;
-  polyfills?: Polyfills;
-
-  /**
-   * Backward-compatible way to set `polyfills.fetch`.
-   */
-  fetchPolyfill?: Polyfills["fetch"];
-
-  /**
-   * Backward-compatible way to set `polyfills.WebSocket`.
-   */
-  WebSocketPolyfill?: Polyfills["WebSocket"];
-} & (
-  | { publicApiKey: string; authEndpoint?: never }
-  | { publicApiKey?: never; authEndpoint: AuthEndpoint }
-);
 
 export type AuthorizeResponse = {
   token: string;
